Add a better sort methord to .muttrc file
In the pervious version, if a new thread of an old email comes, it will
stay with this old email, where is usually been missed.

In the new version, if a new thread comes, it will be put together with
these old threads, but, at the same time, move the whole threads to the
top.

This strategy not only solve the problem mentioned above, but also help
me to understand the thread better.